[GH-2] implement workaround so no build changes are needed
