+++ test/setup.js
@@ -1,0 +1,5 @@
+window.process = {
+  env: {
+    NODE_ENV: 'development'
+  }
+};
